gateway/requirements.txt
flask~=3.0.3
flask-smorest~=0.44.0
python-dotenv~=1.0.1
requests~=2.32.3